+++ exam_django/asset/models/memory.py
